requirements.txt
beautifulsoup4==4.12.3
billiard==4.2.0
chardet==3.0.4
colorama==0.4.6
decorator==5.1.1
deep-translator==1.11.4
django-cors-headers==4.4.0
django-appconf==1.0.6
django-jazzmin==3.0.0
django-parler==2.3
django-watermark==0.2.0
django-webp-converter==0.3.0
Faker==30.1.0
future==1.0.0
GDAL==2.4.0
geocoder==1.38.1
django-admin-extra-buttons
googletrans==4.0.0rc1
watchdog
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
setuptools
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ratelim==0.1.6
setuptools==75.1.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
beautifulsoup4==4.12.3
lxml==5.3.0
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.16
django-leaflet==0.31.0
django-admin-geomap==1.6
django-colorfield==0.11.0
django-environ==0.11.2
django-filter==24.2
django-hashid-field==3.4.1
django-js-asset==2.2.0
django-mptt==0.14.0
django-resized==1.0.2
django-redis==5.4.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-spectacular==0.27.2
drf-writable-nested==0.7.0
gunicorn==23.0.0
hashids==1.3.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
package_name==0.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.5
pycparser==2.22
python-dateutil==2.9.0.post0
python-magic==0.4.27
PyYAML==6.0.2
redis==4.4.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.5.1
tinycss2==1.3.0
tzdata==2024.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
django-debug-toolbar==4.4.6
django-cachalot==2.7.0
channels==4.0.0
channels_redis==4.0.0
daphne==4.0.0